Move NotFound from /app to /app/[domain] and extract header data loading to ServerSideHeader component
