+++ site/src/app/[domain]/[language]/layout.tsx
@@ -1,10 +1,6 @@
-import { gql, previewParams } from "@comet/cms-site";
-import { GQLLayoutQuery, GQLLayoutQueryVariables } from "@src/app/[domain]/[language]/layout.generated";
 import { IntlProvider } from "@src/common/intl/IntlProvider";
 import { loadMessages } from "@src/common/intl/loadMessages";
-import { Header } from "@src/layout/header/Header";
-import { headerFragment } from "@src/layout/header/Header.fragment";
-import { createGraphQLFetch } from "@src/util/graphQLClient";
+import ServerSideHeader from "@src/layout/header/ServerSideHeader";
 import { getSiteConfigForDomain } from "@src/util/siteConfig";
 import { notFound } from "next/navigation";
 import { PropsWithChildren } from "react";
@@ -18,24 +14,9 @@ export default async function Page({ children, params: { language, domain } }: P
 
     const messages = await loadMessages(language);
 
-    const { previewData } = (await previewParams()) || { previewData: undefined };
-    const graphqlFetch = createGraphQLFetch(previewData);
-
-    const { header } = await graphqlFetch<GQLLayoutQuery, GQLLayoutQueryVariables>(
-        gql`
-            query Layout($domain: String!, $language: String!) {
-                header: mainMenu(scope: { domain: $domain, language: $language }) {
-                    ...Header
-                }
-            }
-            ${headerFragment}
-        `,
-        { domain, language },
-    );
-
     return (
         <IntlProvider locale={language} messages={messages}>
-            <Header header={header} />
+            <ServerSideHeader domain={domain} language={language} />
             {children}
         </IntlProvider>
     );
